+++ documentation/ChangeLog.md
@@ -1,6 +1,9 @@
 ## release ChangeLog
 
 ### (NextVersion)
+ * Added support for source formatting using `dfmt` (`Ctrl+Shift+F`). #TODO doc
+   * Added `Format DUB package (dfmt)` command to Project context menu.
+   * Added "Format automatically on editor save." option.
  * Added signing to releases.